Abstract the task run menu into its own component
