+++ src/components/index.ts
@@ -255,6 +255,7 @@ export { default as TaskRunIconText } from './TaskRunIconText.vue'
 export { default as TaskRunList } from './TaskRunList.vue'
 export { default as TaskRunListItem } from './TaskRunListItem.vue'
 export { default as TaskRunLogs } from './TaskRunLogs.vue'
+export { default as TaskRunMenu } from './TaskRunMenu.vue'
 export { default as TaskRunsDeleteButton } from './TaskRunsDeleteButton.vue'
 export { default as TaskRunsSort } from './TaskRunsSort.vue'
 export { default as TimeSpanFilter } from './TimeSpanFilter.vue'
